+++ src/admin/listeners.js
@@ -128,6 +128,7 @@ export const onChannelUpdate = async (oldChannel, newChannel) => {
   const perm = chnUp.permissionOverwrites;
 
   //basic operations
+  if (process.env.DEBUG === "no" && isTestServer(newChannel)) return; //if in prod && modif in test server
   const logChannel = await getLogChannel(newChannel); //get logChannelId
   if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
   const embed = setupEmbed("DarkAqua", chnUp, newChannel, "tag"); //setup embed
@@ -312,9 +313,9 @@ export const onThreadCreate = async (thread, newly) => {
   if (thread) {
     //sometimes thread is null
     if (thread.joinable && !thread.joined) await thread.join(); //join thread created
-    const logChannel = await getLogChannel(thread); //get logChannel
-    if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
+    if (process.env.DEBUG === "no" && isTestServer(thread)) return; //if in prod && modif in test server
 
+    const logChannel = await getLogChannel(thread); //get logChannel
     const perso = PERSONALITY.getAdmin().threadCreate;
     const log = await fetchAuditLog(thread.guild, "ThreadCreate", 1); //get auditLog
     const executor = log.executor
@@ -366,6 +367,7 @@ export const onRoleUpdate = async (oldRole, newRole) => {
   const roleUp = personality.roleUpdate;
   const auditLog = personality.auditLog;
 
+  if (process.env.DEBUG === "no" && isTestServer(newRole)) return; //if in prod && modif in test server
   const logChannel = await getLogChannel(newRole); //get logChannelId
   if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
   const embed = setupEmbed("DarkGold", roleUp, newRole); //setup embed
@@ -456,8 +458,8 @@ export const onMessageDelete = async (message) => {
   const messageDel = personality.messageDelete;
   const auditLog = personality.auditLog;
 
+  if (process.env.DEBUG === "no" && isTestServer(message)) return; //if in prod && modif in test server
   const logChannel = await getLogChannel(message, "thread"); //get logChannel
-  if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
 
   const uDate = new Date(message.createdAt); //set date as Date object
   if (currentServer.name === "prod") uDate.setHours(uDate.getHours() + 1); //add 1h to date
@@ -599,8 +601,8 @@ export const onMessageUpdate = async (oldMessage, newMessage) => {
   const messageU = personality.messageUpdate;
   const auditLog = personality.auditLog;
 
+  if (process.env.DEBUG === "no" && isTestServer(newMessage)) return; //if in prod && modif in test server
   const logChannel = await getLogChannel(nMessage, "thread"); //get logChannel
-  if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
 
   const embed = setupEmbed("DarkGreen", messageU, nMessage.author, "tag"); //setup embed
   //no auditLog when message update
@@ -804,8 +806,8 @@ export const onGuildMemberUpdate = async (oldMember, newMember) => {
   const timeout = personality.timeout;
   const auditLog = personality.auditLog;
 
+  if (process.env.DEBUG === "no" && isTestServer(newMember)) return; //if in prod && modif in test server
   const logChannel = await getLogChannel(newMember); //get logChannel
-  if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
   const embed = setupEmbed("Orange", timeout, user, "tag"); //setup embed
   const timeoutLog = await fetchAuditLog(newMember.guild, "MemberUpdate", 1); //get auditLog
   const reason = timeoutLog.reason; //get ban reason
@@ -832,8 +834,7 @@ export const onGuildMemberRemove = async (memberKick) => {
   const personality = PERSONALITY.getAdmin(); //get personality
   const auditLog = personality.auditLog;
 
-  const logChannel = await getLogChannel(memberKick); //get logChannel
-  if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
+  if (process.env.DEBUG === "no" && isTestServer(memberKick)) return; //if in prod && modif in test server
   const kickLog = await fetchAuditLog(memberKick.guild, "MemberKick", 1); //get auditLog
   const reason = kickLog ? kickLog.reason : null; //get kick reason
 
@@ -855,12 +856,14 @@ export const onGuildMemberRemove = async (memberKick) => {
     //no log or too old => not kicked but left
     const guildKick = personality.guildKick.leave;
     const embed = setupEmbed("DarkPurple", guildKick, userKick, "user"); //setup embed
+    embed.addFields({ name: guildKick.id, value: memberKick.id, inline: true }); //add user id
     if (textRoles)
       embed.addFields({
         name: guildKick.roles,
         value: textRoles,
         inline: true,
       }); //add user roles if any
+    const logChannel = await getLogChannel(memberKick, "inAndOut"); //get logChannel
     const messageList = await endCasesEmbed(
       userKick,
       kickLog,
@@ -878,8 +881,10 @@ export const onGuildMemberRemove = async (memberKick) => {
 
   const guildKick = personality.guildKick.kick;
   const embed = setupEmbed("DarkPurple", guildKick, userKick, "user"); //setup embed
+  embed.addFields({ name: guildKick.id, value: memberKick.id, inline: true }); //add user id
   if (textRoles)
     embed.addFields({ name: guildKick.roles, value: textRoles, inline: true }); //add user roles if any
+  const logChannel = await getLogChannel(memberKick); //get logChannel
 
   endCasesEmbed(
     userKick,